requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.14.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.34.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.2
pandas==2.2.1
pillow==10.4.0
plotly==5.20.0
protobuf==4.25.6
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.32.2
sympy==1.13.3
tenacity==8.5.0
textblob==0.18.0.post0
threadpoolctl==3.6.0
tokenizers==0.13.3
toml==0.10.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.28.1
typing_extensions==4.13.1
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.3.0
wordcloud==1.9.4
